+++ packages/ci/src/lib/cli/persist.ts
@@ -7,11 +7,14 @@ import {
   DEFAULT_PERSIST_OUTPUT_DIR,
   type Format,
   persistConfigSchema,
+  uploadConfigSchema,
 } from '@code-pushup/models';
-import { objectFromEntries, stringifyError } from '@code-pushup/utils';
+import { createReportPath, objectFromEntries } from '@code-pushup/utils';
+
+export type EnhancedPersistConfig = Pick<CoreConfig, 'persist' | 'upload'>;
 
 export function persistedFilesFromConfig(
-  config: Pick<CoreConfig, 'persist'>,
+  config: EnhancedPersistConfig,
   { isDiff, directory }: { isDiff?: boolean; directory: string },
 ): Record<Format, string> {
   const {
@@ -24,23 +27,28 @@ export function persistedFilesFromConfig(
   const dir = path.isAbsolute(outputDir)
     ? outputDir
     : path.join(directory, outputDir);
-  const name = isDiff ? `${filename}-diff` : filename;
+  const suffix = isDiff ? 'diff' : undefined;
 
   return objectFromEntries(
     DEFAULT_PERSIST_FORMAT.map(format => [
       format,
-      path.join(dir, `${name}.${format}`),
+      createReportPath({ outputDir: dir, filename, format, suffix }),
     ]),
   );
 }
 
 export async function parsePersistConfig(
   json: unknown,
-): Promise<Pick<CoreConfig, 'persist'>> {
-  const schema = z.object({ persist: persistConfigSchema.optional() });
+): Promise<EnhancedPersistConfig> {
+  const schema = z.object({
+    persist: persistConfigSchema.optional(),
+    upload: uploadConfigSchema.optional(),
+  });
   const result = await schema.safeParseAsync(json);
   if (result.error) {
-    throw new Error(`Invalid persist config - ${stringifyError(result.error)}`);
+    throw new Error(
+      `Code PushUp config is invalid:\n${z.prettifyError(result.error)}`,
+    );
   }
   return result.data;
 }