+++ packages/ci/src/lib/constants.ts
@@ -14,4 +14,9 @@ export const DEFAULT_SETTINGS: Settings = {
   logger: console,
   nxProjectsFilter: '--with-target={task}',
   skipComment: false,
+  configPatterns: null,
+  searchCommits: false,
 };
+
+export const MIN_SEARCH_COMMITS = 1;
+export const MAX_SEARCH_COMMITS = 100;