+++ packages/cli/src/lib/implementation/compare.model.ts
@@ -1,3 +1,0 @@
-import type { Diff } from '@code-pushup/utils';
-
-export type CompareOptions = Diff<string> & { label?: string };