+++ packages/models/src/lib/audit.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
+import { createDuplicateSlugsCheck } from './implementation/checks.js';
 import { metaSchema, slugSchema } from './implementation/schemas.js';
-import { errorItems, hasDuplicateStrings } from './implementation/utils.js';
 
 export const auditSchema = z
   .object({
@@ -18,28 +18,7 @@ export const auditSchema = z
 
 export type Audit = z.infer<typeof auditSchema>;
 export const pluginAuditsSchema = z
-  .array(auditSchema, {
-    description: 'List of audits maintained in a plugin',
-  })
+  .array(auditSchema)
   .min(1)
-  // audit slugs are unique
-  .refine(
-    auditMetadata => !getDuplicateSlugsInAudits(auditMetadata),
-    auditMetadata => ({
-      message: duplicateSlugsInAuditsErrorMsg(auditMetadata),
-    }),
-  );
-
-// =======================
-
-// helper for validator: audit slugs are unique
-function duplicateSlugsInAuditsErrorMsg(audits: Audit[]) {
-  const duplicateRefs = getDuplicateSlugsInAudits(audits);
-  return `In plugin audits the following slugs are not unique: ${errorItems(
-    duplicateRefs,
-  )}`;
-}
-
-function getDuplicateSlugsInAudits(audits: Audit[]) {
-  return hasDuplicateStrings(audits.map(({ slug }) => slug));
-}
+  .check(createDuplicateSlugsCheck('Audit'))
+  .describe('List of audits maintained in a plugin');