+++ CHANGELOG.md
@@ -1,3 +1,88 @@
+## 0.73.0 (2025-08-08)
+
+### 🚀 Features
+
+- **ci:** add searchCommits option to extend portal cache range ([0b14111f](https://github.com/code-pushup/cli/commit/0b14111f))
+
+### ❤️ Thank You
+
+- Matěj Chalk
+
+## 0.72.1 (2025-08-06)
+
+### 🩹 Fixes
+
+- **ci:** download portal report for base commit and include details ([21572d52](https://github.com/code-pushup/cli/commit/21572d52))
+
+### ❤️ Thank You
+
+- Matěj Chalk
+
+## 0.72.0 (2025-08-06)
+
+### 🚀 Features
+
+- **ci:** add configPatterns as optional performance optimization ([823ade1f](https://github.com/code-pushup/cli/commit/823ade1f))
+- **cli:** use default report paths if --before/--after missing in compare command ([61ee2272](https://github.com/code-pushup/cli/commit/61ee2272))
+- **core:** copy label from report.json inputs to report-diff.json output ([2adcf9db](https://github.com/code-pushup/cli/commit/2adcf9db))
+- **models:** add optional label to report.json schema ([77d33b54](https://github.com/code-pushup/cli/commit/77d33b54))
+- **utils:** interpolate variables in strings ([242435e0](https://github.com/code-pushup/cli/commit/242435e0))
+
+### 🔥 Performance
+
+- **ci:** use bulk command to compare reports for all projects ([f9deac78](https://github.com/code-pushup/cli/commit/f9deac78))
+
+### ❤️ Thank You
+
+- Matěj Chalk
+
+## 0.71.0 (2025-08-01)
+
+### 🚀 Features
+
+- **ci:** download report from graphql api and convert to report.json format ([d35cd7a0](https://github.com/code-pushup/cli/commit/d35cd7a0))
+- **ci:** parse upload along with persist in print-config commands ([a73bf21b](https://github.com/code-pushup/cli/commit/a73bf21b))
+- **ci:** download previous report from portal if available ([918eb0d5](https://github.com/code-pushup/cli/commit/918eb0d5))
+- **utils:** add type helpers for lower/upper case conversions ([50483748](https://github.com/code-pushup/cli/commit/50483748))
+
+### ❤️ Thank You
+
+- Matěj Chalk
+
+## 0.70.0 (2025-07-31)
+
+### 🚀 Features
+
+- **ci:** update to zod v4 ([c342dab1](https://github.com/code-pushup/cli/commit/c342dab1))
+- **ci:** remove redundant --no-progress argument ([b0ed2fa9](https://github.com/code-pushup/cli/commit/b0ed2fa9))
+- **ci:** skip persist.format args if defaults already configured ([c11db95e](https://github.com/code-pushup/cli/commit/c11db95e))
+- **cli:** disable progress bar in CI environment ([ecd42c60](https://github.com/code-pushup/cli/commit/ecd42c60))
+- **models:** add generic artifact generation to enable caching ([#1023](https://github.com/code-pushup/cli/pull/1023))
+- **models:** update to zod v4 ([7985c239](https://github.com/code-pushup/cli/commit/7985c239))
+- **nx-plugin:** update to zod v4 ([91774182](https://github.com/code-pushup/cli/commit/91774182))
+- **plugin-coverage:** update to zod v4 ([a0814d31](https://github.com/code-pushup/cli/commit/a0814d31))
+- **plugin-eslint:** update zod to v4 ([4ef657f1](https://github.com/code-pushup/cli/commit/4ef657f1))
+- **plugin-js-packages:** update to zod v4 ([c24b7435](https://github.com/code-pushup/cli/commit/c24b7435))
+- **plugin-jsdocs:** update to zod v4 ([3444e8d3](https://github.com/code-pushup/cli/commit/3444e8d3))
+- **plugin-lighthouse:** implement multiple URL support ([f4db1368](https://github.com/code-pushup/cli/commit/f4db1368))
+- **plugin-typescript:** update to zod v4 ([6de416f4](https://github.com/code-pushup/cli/commit/6de416f4))
+- **utils:** update to zod v4, replace zod-validation-error with z.prettifyError ([b1364d7d](https://github.com/code-pushup/cli/commit/b1364d7d))
+- **utils:** parse boolean environment variables ([be08c188](https://github.com/code-pushup/cli/commit/be08c188))
+- **utils:** detect ci environment ([3430d31b](https://github.com/code-pushup/cli/commit/3430d31b))
+- **utils:** extract boolean value coercion to standalone function ([f76e4fb0](https://github.com/code-pushup/cli/commit/f76e4fb0))
+
+### 🩹 Fixes
+
+- **ci:** do not set --verbose by default ([ab39b050](https://github.com/code-pushup/cli/commit/ab39b050))
+- **cli:** adapt format schema check to zod v4 ([77a384b0](https://github.com/code-pushup/cli/commit/77a384b0))
+- **models:** use implementAsync for async z.function occurrences ([17931228](https://github.com/code-pushup/cli/commit/17931228))
+
+### ❤️ Thank You
+
+- Andrii Siuta
+- Hanna Skryl @hanna-skryl
+- Matěj Chalk
+
 ## 0.69.5 (2025-07-11)
 
 ### 🩹 Fixes